Fix signed/unsigned warning
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14952 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/mpegplayer/alloc.c
@@ -45,7 +45,7 @@ static void * mpeg_malloc_internal (unsigned char *mallocbuf,
 {
     void *x;
 
-    if (*mem_ptr + size > bufsize)
+    if ((size_t) (*mem_ptr + size) > bufsize)
     {
         DEBUGF("OUT OF MEMORY\n");
         return NULL;
